backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==4.0.1
python-multipart==0.0.9
python-dotenv==1.0.1
pandas==2.2.2
openpyxl==3.1.2
pydantic[email]==2.7.1
pydantic-settings==2.2.1
